% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SupportFun.R
\name{crossover_and_mutate}
\alias{crossover_and_mutate}
\title{Perform Crossover and Mutation}
\usage{
crossover_and_mutate(
  parent1,
  parent2,
  lr_data,
  mutation_rate = 0.1,
  lr_mutation_boost = 0.3
)
}
\arguments{
\item{parent1}{A binary vector representing the first parent.}

\item{parent2}{A binary vector representing the second parent.}

\item{lr_data}{A data frame containing ligand-receptor pairs.}

\item{mutation_rate}{The probability of mutation.}

\item{lr_mutation_boost}{The increased mutation probability for ligand-receptor genes.}
}
\value{
A binary vector representing the offspring.
}
\description{
This function performs crossover and mutation to generate offspring.
}
