% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SupportFun.R
\name{evaluate}
\alias{evaluate}
\title{Evaluate Fitness of a Solution}
\usage{
evaluate(features, original_data, lr_data, deg_results, prune_rate)
}
\arguments{
\item{features}{A binary vector indicating which features are active.}

\item{original_data}{The original data frame containing gene expression data and labels.}

\item{lr_data}{A data frame containing ligand-receptor pairs.}

\item{deg_results}{A data frame containing differential expression results.}

\item{prune_rate}{A numeric value indicating the proportion of features to prune.}
}
\value{
A list containing fitness, train error, Q_LR, logFC, and selected features.
}
\description{
This function evaluates the fitness of a solution based on classification error, ligand-receptor interactions, and differential expression.
}
