% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SupportFun.R
\name{generate_population}
\alias{generate_population}
\title{Generate Initial Population}
\usage{
generate_population(pop_size, num_features, prob = 0.1)
}
\arguments{
\item{pop_size}{The size of the population.}

\item{num_features}{The number of features in the data.}

\item{prob}{The probability of a feature being active in the initial population.}
}
\value{
A matrix representing the initial population.
}
\description{
This function generates an initial population of solutions for the genetic algorithm.
}
