% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SupportFun.R
\name{preprocess_data}
\alias{preprocess_data}
\title{Preprocess Data for Genetic Algorithm}
\usage{
preprocess_data(data, labels, variance_threshold = 0.8)
}
\arguments{
\item{data}{A matrix or data frame of gene expression data.}

\item{labels}{A vector of labels for each sample.}

\item{variance_threshold}{A numeric value between 0 and 1 indicating the proportion of genes to retain based on variance.}
}
\value{
A list containing the filtered data and DEG results.
}
\description{
This function preprocesses the input data by filtering genes based on variance and performing differential expression analysis.
}
