% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thca_gse_154763.R
\docType{data}
\name{thca_gse_154763}
\alias{thca_gse_154763}
\title{THCA GSE154763 Dataset}
\format{
A data frame with gene expression values.
}
\source{
GSE154763 dataset from GEO database.
}
\usage{
thca_gse_154763
}
\description{
This dataset contains gene expression data for thyroid cancer.
}
\examples{
data(thca_gse_154763)
}
\keyword{datasets}
