% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thca_gse_154763_Labels.R
\docType{data}
\name{thca_gse_154763_Labels}
\alias{thca_gse_154763_Labels}
\title{THCA GSE154763 Labels}
\format{
A vector indicating the condition of each sample (e.g., tumor or normal).
}
\source{
GSE154763 dataset from GEO database.
}
\usage{
thca_gse_154763_Labels
}
\description{
This dataset contains labels for the thyroid cancer gene expression data.
}
\examples{
data(thca_gse_154763_Labels)
}
\keyword{datasets}
